Use the dracula pro theme and remove i3status-rust
Changed back to use bumblebee-status instead of i3status-rust
since I prefer it in functionality. i3status-rust still has
an issue open where I've been waiting a while for it to be fixed
which was not done as of yet.

Since I like the dracula theme, I changed the themes to match dracula.